Make -x -i options work on all filenames
The include option was filtering based on mark'd mounted filesystems
which meant it don't do what the user expected.  Instead mark all
mounted file systems and filter based on what is being notified.

Make the -x option take higher precedence over the -i option and
note that in the manual

Signed-off-by: Colin Ian King <colin.king@canonical.com>