fix memory leak when using the -n option
The ‘-n’ sets OPT_NOSTATS | OPT_VERBOSE, OPT_NOSTATS being set means that the
file_stat_dump is not called and memory is not free'd.  Fix this by calling
file_stat_free(). Also nullify file_stats entries when they are free'd.

Kudos to Michael Glad for reporting and analysing the root cause of this
issue.

Signed-off-by: Colin Ian King <colin.king@canonical.com>